Skeleton is now shown on image loading on both settings and finidng edit.

+++ frontend/app/routes/spaces.$spaceName.jsx
@@ -51,6 +51,19 @@ export async function action({request, params}){
         })
     }
 
+    if(request.method ==='POST'){
+        const url = res.get('imageUrl')
+        
+        return prisma.finding.update({
+            where: {
+                id : Number(res.get('findingId')),
+            },
+            data : {
+                image: url
+            }
+        })
+    }
+
     if(request.method === 'DELETE'){
         return await prisma.space.delete({
             where: {
